+++ src/paimon/common/utils/var_length_int_utils.h
@@ -88,13 +88,13 @@ class VarLengthIntUtils {
             return static_cast<int32_t>(first_byte);
         }
         // Multi-byte: fall through to generic loop
-        int32_t result = 0;
+        uint32_t result = 0;
         for (int32_t shift = 0; shift < 32; shift += 7) {
             auto byte_val = static_cast<uint8_t>(data[*offset]);
             ++(*offset);
-            result |= static_cast<int32_t>(byte_val & 0x7F) << shift;
+            result |= static_cast<uint32_t>(byte_val & 0x7F) << shift;
             if ((byte_val & 0x80) == 0) {
-                return result;
+                return static_cast<int32_t>(result);
             }
         }
         return Status::Invalid("Malformed varint32: too many continuation bytes");
@@ -109,13 +109,13 @@ class VarLengthIntUtils {
             return static_cast<int64_t>(first_byte);
         }
         // Multi-byte: fall through to generic loop
-        int64_t result = 0;
+        uint64_t result = 0;
         for (int32_t shift = 0; shift < 64; shift += 7) {
             auto byte_val = static_cast<uint8_t>(data[*offset]);
             ++(*offset);
-            result |= static_cast<int64_t>(byte_val & 0x7F) << shift;
+            result |= static_cast<uint64_t>(byte_val & 0x7F) << shift;
             if ((byte_val & 0x80) == 0) {
-                return result;
+                return static_cast<int64_t>(result);
             }
         }
         return Status::Invalid("Malformed varint64: too many continuation bytes");
